+++ doc/release-notes.rst
@@ -31,7 +31,7 @@ In order to ensure compatibility with the new code, please delete these sections
   * ``powerplantmatching.heuristics.set_denmark_region_id``
   * ``powerplantmatching.heuristics.remove_oversea_areas``
   * ``powerplantmatching.heuristics.set_known_retire_years``
-
+* The argument ``extendby_kwargs`` in the function ``powerplantmatching.collection.matched_data`` was deprecated in the favor of ``extend_by_kwargs``.
 
 
 **Non-Breaking Code Changes**
@@ -48,6 +48,9 @@ In order to ensure compatibility with the new code, please delete these sections
 * The `GEO` data now returns a dataset containing power plant units.  
 * The ``ESE`` dataset was removed due the hosting website taken down. 
 * The argument ``subsume_uncommon_fueltypes_to_other`` in ``powerplantmatching.collection.matched_data`` was removed. 
+* The function ``powerplantmatching.cleaning.aggregate_units`` does not support the arguments `use_saved_aggregation` and `save_aggregation` anymore due to it's unsecure behavior.
+* The function ``powerplantmatching.matching.compare_two_datasets`` does not support the arguments `use_saved_matches` anymore due to it's unsecure behavior.
+
 
 Version 0.4.6 (25.11.2020)
 --------------------------